Split ml and data to another module
